DRY principle enforced to allow for greater code flexibility.
To be precise, using a map to store all available commands and their string descriptions removes the nessesity for a redundant switch statement that has to be manually updated
with every new command. As a result, the run function is now more clean and concise.
